tools/kaitu-signer/requirements.txt
pystray>=0.19
pywinauto>=0.6.8
boto3>=1.34
pyotp>=2.9
Pillow>=10.0
